correcion de validacion de registro
