+++ src/views/user/register.ejs
@@ -15,17 +15,17 @@
                 <label for="nombre">Nombre</label>
                 <input type="text" id="nombre" name="nombre" placeholder="Nombre" value="<%= locals.old && old.nombre ? old.nombre : ''%>" required>
                 <label for="apellido">Apellido</label>
-                <input type="text" id="apellido" name="apellido" placeholder="Apellido" value="<%= locals.old && old.apellido ? old.apellido : ''%>" required>
+                <input type="text" id="apellido" name="apellido" placeholder="Apellido" value="<%= locals.old && old.apellido ? old.apellido : ''%>"  required>
                 <label for="correo">Correo Electrónico</label>
                 <input type="email" id="correo" name="correo" placeholder="Correo Electrónico" value="<%= locals.old && old.correo ? old.correo : ''%>" required>
                 <label for="usuario">Nombre de Usuario</label>
-                <input type="text" id="usuario" name="usuario" placeholder="Nombre de Usuario" value="<%= locals.old && old.usuario ? old.usuario : ''%>" required>
+                <input type="text" id="usuario" name="usuario" placeholder="Nombre de Usuario" value="<%= locals.old && old.usuario ? old.usuario : ''%>"  required>
                 <label for="contrasena">Contraseña</label>
-                <input type="password" id="contrasena" name="contrasena" placeholder="Contraseña">
+                <input type="password" id="contrasena" name="contrasena" placeholder="Contraseña" required>
                 <label for="confirmar_contrasena">Confirmar Contraseña</label>
-                <input type="password" id="confirmar_contrasena" name="confirmar_contrasena" placeholder="Confirmar Contraseña">
+                <input type="password" id="confirmar_contrasena" name="confirmar_contrasena" placeholder="Confirmar Contraseña" required>
                 <label for="imagen">Imagen de usuario:</label>
-                <input type="file" id="imagen" name="imagen" accept="image/*">
+                <input type="file" id="imagen" name="imagen" accept="image/*" required>
                 <button class="btn" type="submit">Registrarse</button>
             </form>
         </section>
